Move `run_client_app` to `supernode.app` (#3352)
